fixed failed build because of old tests
